# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is sample node js swagger api.",
    "version": "1.0.0",
    "title": "Swagger Uber"
  },
...
  "host": "localhost:3900",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "paths": {
    "/users": {
      "post": {
        "tags": ["User"],
        "summary": "Add a new user",
        "description": "",
        "operationId": "addUser",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": ["User"],
        "summary": "Returns user",
        "description": "Get all users",
        "operationId": "getUser",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/login": {
      "post": {
        "tags": ["User"],
        "summary": "Login a user",
        "description": "",
        "operationId": "loginUser",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/cabs": {
      "post": {
        "tags": ["Cab"],
        "summary": "Create a cab",
        "description": "",
        "operationId": "createCab",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Cab object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Cab"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": ["Cab"],
        "summary": "Returns cabs",
        "description": "Returns a cab",
        "operationId": "getCab",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/history": {
      "get": {
        "tags": ["Cab History"],
        "summary": "Get all user cab history",
        "description": "Retrieves all user cab history.",
        "operationId": "getHistory",
        "produces": ["application/xml", "application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/History"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "fullName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "accountStatus": {
          "type": "boolean"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "Login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "Cab": {
      "type": "object",
      "properties": {
        "cabName": {
          "type": "string"
        },
        "cabType": {
          "type": "string"
        },
        "cabModel": {
          "type": "string"
        },
        "cabLat": {
          "type": "string"
        },
        "canLon": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "area": {
          "type": "string"
        },
        "status": {
          "type": "boolean"
        },
        "driverId": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Cab"
      }
    },
    "History": {
      "type": "object",
      "properties": {
        "price": {
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "cabId": {
          "type": "string"
        }
      },
      "xml": {
        "name": "History"
      }
    }
  }
}